// miniprogram/api/generated/models/SearchUserItemDto.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SearchUserItemDto
 */
export interface SearchUserItemDto {
    /**
     * 
     * @type {number}
     * @memberof SearchUserItemDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof SearchUserItemDto
     */
    nickname?: string;
    /**
     * 
     * @type {string}
     * @memberof SearchUserItemDto
     */
    avatarUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof SearchUserItemDto
     */
    bio?: string;
    /**
     * 
     * @type {number}
     * @memberof SearchUserItemDto
     */
    articleCount?: number;
    /**
     * 
     * @type {number}
     * @memberof SearchUserItemDto
     */
    likeCount?: number;
}

/**
 * Check if a given object implements the SearchUserItemDto interface.
 */
export function instanceOfSearchUserItemDto(value: object): value is SearchUserItemDto {
    return true;
}

export function SearchUserItemDtoFromJSON(json: any): SearchUserItemDto {
    return SearchUserItemDtoFromJSONTyped(json, false);
}

export function SearchUserItemDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): SearchUserItemDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'nickname': json['nickname'] == null ? undefined : json['nickname'],
        'avatarUrl': json['avatar_url'] == null ? undefined : json['avatar_url'],
        'bio': json['bio'] == null ? undefined : json['bio'],
        'articleCount': json['article_count'] == null ? undefined : json['article_count'],
        'likeCount': json['like_count'] == null ? undefined : json['like_count'],
    };
}

export function SearchUserItemDtoToJSON(json: any): SearchUserItemDto {
    return SearchUserItemDtoToJSONTyped(json, false);
}

export function SearchUserItemDtoToJSONTyped(value?: SearchUserItemDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'nickname': value['nickname'],
        'avatar_url': value['avatarUrl'],
        'bio': value['bio'],
        'article_count': value['articleCount'],
        'like_count': value['likeCount'],
    };
}