// miniprogram/api/generated/models/TopicItemDto.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TopicItemDto
 */
export interface TopicItemDto {
    /**
     * 
     * @type {number}
     * @memberof TopicItemDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof TopicItemDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof TopicItemDto
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof TopicItemDto
     */
    coverUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof TopicItemDto
     */
    status?: string;
    /**
     * 
     * @type {number}
     * @memberof TopicItemDto
     */
    articleCount?: number;
    /**
     * 
     * @type {number}
     * @memberof TopicItemDto
     */
    viewCount?: number;
    /**
     * 
     * @type {string}
     * @memberof TopicItemDto
     */
    startAt?: string;
    /**
     * 
     * @type {string}
     * @memberof TopicItemDto
     */
    endAt?: string;
    /**
     * 
     * @type {number}
     * @memberof TopicItemDto
     */
    daysRemaining?: number;
}

/**
 * Check if a given object implements the TopicItemDto interface.
 */
export function instanceOfTopicItemDto(value: object): value is TopicItemDto {
    return true;
}

export function TopicItemDtoFromJSON(json: any): TopicItemDto {
    return TopicItemDtoFromJSONTyped(json, false);
}

export function TopicItemDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TopicItemDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'] == null ? undefined : json['title'],
        'description': json['description'] == null ? undefined : json['description'],
        'coverUrl': json['cover_url'] == null ? undefined : json['cover_url'],
        'status': json['status'] == null ? undefined : json['status'],
        'articleCount': json['article_count'] == null ? undefined : json['article_count'],
        'viewCount': json['view_count'] == null ? undefined : json['view_count'],
        'startAt': json['start_at'] == null ? undefined : json['start_at'],
        'endAt': json['end_at'] == null ? undefined : json['end_at'],
        'daysRemaining': json['days_remaining'] == null ? undefined : json['days_remaining'],
    };
}

export function TopicItemDtoToJSON(json: any): TopicItemDto {
    return TopicItemDtoToJSONTyped(json, false);
}

export function TopicItemDtoToJSONTyped(value?: TopicItemDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'description': value['description'],
        'cover_url': value['coverUrl'],
        'status': value['status'],
        'article_count': value['articleCount'],
        'view_count': value['viewCount'],
        'start_at': value['startAt'],
        'end_at': value['endAt'],
        'days_remaining': value['daysRemaining'],
    };
}